Support milliseconds in cacheDuration parsing
Since the introduction of cacheDuration parsing in 25cbddd we've been
seeing parsing failures for one of our IdPs, whose cacheDuration value
is set to

    cacheDuration="PT6H0M0.000S"

This seems like a perfectly valid ISO8601 duration - the regexp being
used for parsing doesn't provide for the possibility of milliseconds.

This adds an optional non-captured group to the regexp to permit them.
Since the captured string value is already being converted using .to_f,
it appears this should work as-is.

To achieve minimal test coverage I've added a seconds string to one of
the existing metadata examples with a zero value; this reproduces the
failure and confirms the fix without needing to add a specific test
case.